move tests to/from test_indexing.py
